third_party/blink/web_tests/platform/generic/virtual/text-antialias/soft-hyphen-5-expected.txt

Soft Hyphen Test

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS true is true

anti­dis­est­ab­lish­ment­arian­ism.
